refactor: Use new folder structure in frontend
