+++ src/frontend/plugins/index.tsx
@@ -1,26 +1,26 @@
-import { createAnchorPlugin } from '@edtr-io/plugin-anchor'
-import { createGeogebraPlugin } from '@edtr-io/plugin-geogebra'
-import { createHighlightPlugin } from '@edtr-io/plugin-highlight'
+import { createAnchorPlugin } from '@frontend/src/serlo-editor/plugins/anchor'
+import { createGeogebraPlugin } from '@frontend/src/serlo-editor/plugins/geogebra'
+import { createHighlightPlugin } from '@frontend/src/serlo-editor/plugins/highlight'
 import {
   createInputExercisePlugin,
   InputExerciseType,
-} from '@edtr-io/plugin-input-exercise'
-import { createMultimediaExplanationPlugin } from '@edtr-io/plugin-multimedia-explanation'
-import { createRowsPlugin } from '@edtr-io/plugin-rows'
-import { createScMcExercisePlugin } from '@edtr-io/plugin-sc-mc-exercise'
+} from '@frontend/src/serlo-editor/plugins/input-exercise'
+import { createMultimediaExplanationPlugin } from '@frontend/src/serlo-editor/plugins/multimedia-explanation'
+import { createRowsPlugin } from '@frontend/src/serlo-editor/plugins/rows'
+import { createScMcExercisePlugin } from '@frontend/src/serlo-editor/plugins/sc-mc-exercise'
 import { createSerloInjectionPlugin } from './serlo-injection'
-import { createSpoilerPlugin } from '@edtr-io/plugin-spoiler'
-import { createTextPlugin } from '@edtr-io/plugin-text'
-import { createBoxPlugin } from '@frontend/src/edtr-io/plugins/box'
+import { createSpoilerPlugin } from '@frontend/src/serlo-editor/plugins/spoiler'
+import { createTextPlugin } from '@frontend/src/serlo-editor/plugins/text'
+import { createBoxPlugin } from '@frontend/src/serlo-editor/plugins/box'
+import { equationsPlugin } from '@frontend/src/serlo-editor/plugins/equations'
+import { createSerloTablePlugin } from '@frontend/src/serlo-editor/plugins/serlo-table'
+import { loggedInData } from '@frontend/src/data/de'
 
 import {
   EdusharingConfig,
   createEdusharingAssetPlugin,
 } from './edusharing-asset'
-import { equationsPlugin } from '@frontend/src/edtr-io/plugins/equations'
-import { createSerloTablePlugin } from '@frontend/src/edtr-io/plugins/serlo-table'
 import { registry, getPluginRegistry } from './registry'
-import { loggedInData } from '@frontend/src/data/de'
 
 export function createPlugins(config: EdusharingConfig) {
   return {
