+++ src/frontend/plugins/serlo-injection/index.ts
@@ -3,17 +3,13 @@ import {
   EditorPluginProps,
   string,
   StringStateType,
-} from '@edtr-io/plugin'
+} from '@frontend/src/serlo-editor/plugin'
+
 import { SerloInjectionEditor } from './editor'
 
 /**
  * @param config - {@link SerloInjectionConfig | Plugin configuration}
  */
-
-// this plugin is currently only used in the edusharing integration
-// https://github.com/serlo/serlo-editor-for-edusharing
-
-// TODO: move to edusharing integration
 
 export function createSerloInjectionPlugin(
   config: SerloInjectionConfig = {}
