+++ src/pages/_app.tsx
@@ -1,7 +1,9 @@
 import type { AppProps } from 'next/app'
+
+import '@frontend/src/assets-webkit/styles/serlo-tailwind.css'
+
 import 'katex/dist/katex.min.css'
 import '../frontend/styles.css'
-import '@frontend/src/assets-webkit/styles/serlo-tailwind.css'
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />
