+++ src/pages/editor-for-edusharing.tsx
@@ -1,5 +1,6 @@
 import { useState, useEffect } from 'react'
 import { GetServerSideProps } from 'next'
+
 import {
   migrate,
   emptyDocument,
